Love your routines, however the serial read function would wait forever if it tried to communicate with a servo that didn't exist or had become disconnected.
I added timeouts to the serial read call and a ServoTimeout error handler to differentiate between a checksum
error and a timeout.

This also let me write a program that would scan the bus and report the servo addresses in use and any
addresses that have two or more servos.  It assumes that if it gets a checksum, that the most likely cause
is two servos at that address.   I've tested this on my servos and it works as expected.

Cheers,
Brad